refactor(air): split lookup module into generic + Miden-specific halves
Phase C of the lookup-module split. Carves the field-polymorphic core of
the closure-based LogUp machinery out of `air/src/constraints/lookup/`
into a new top-level `air/src/lookup/` module. The old directory keeps
the Miden-specific wiring.

Moves (history preserved via `git mv`):
- `builder.rs`, `message.rs`, `fractions.rs`, `constraint.rs`
- `debug/trace/{mod.rs,builder.rs}`, `debug/validation/{mod.rs,builder.rs}`
- `trace/challenges.rs` → `lookup/challenges.rs`
- `prover.rs` (moved; `build_lookup_fractions` generalized to `<F, EF>`)

New generic files:
- `lookup/mod.rs` — hosts the `LookupAir` trait + re-exports
- `lookup/aux_builder.rs` — `build_logup_aux<A, F, EF>` composing
  collection + accumulation, sized from the caller
- `lookup/debug/mod.rs` — generic re-exports for the debug surface
- `lookup/debug/validation/mod.rs` — split of the old `validate()`:
  generic `validate_structure_only` returns `StructureReport`

Miden-side stays in `constraints/lookup/`:
- `miden_air.rs`, `main_air.rs`, `chiplet_air.rs`, `buses/*`, `bus_id.rs`
- `aux_builder.rs` — `MidenLookupAuxBuilder` now a thin wrapper that
  sources Miden periodic columns + bus shape and delegates to
  `build_logup_aux`
- `extension_impls.rs` — picks up the `Main/ChipletLookupBuilder`
  impls for the generic adapters + debug builders (gated on
  `cfg(test)`/`bus-debug`)
- `debug/validation/mod.rs` — Miden wrapper `validate()` composing
  `validate_structure_only` + `check_symbolic_degrees`, returning the
  full `ValidationReport`
- `debug/validation/symbolic.rs` — unchanged (imports `ProcessorAir`,
  `NUM_PUBLIC_VALUES`, `DEGREE_BUDGET`)

Renames & re-exports:
- `LookupChallenges` → `Challenges` at every call site; the
  `constraints/lookup/challenges.rs` alias file is deleted.
- `air/src/lib.rs` drops the old `pub mod lookup { pub use … }` block
  in favor of `pub mod lookup;`; Miden items (`MidenLookupAir`,
  `MidenLookupAuxBuilder`, `MIDEN_MAX_MESSAGE_WIDTH`, `NUM_BUS_IDS`,
  `ColumnOracleBuilder`) are re-exported from `crate::lookup` for
  ergonomic access.
- `trace::Challenges` reference in `lib.rs`/`ace.rs` → `lookup::Challenges`.
- `crate::trace::challenges` module is deleted.

Verified: `make check`, `make lint`, `cargo test -p miden-air --lib`
(50 pass), `cargo test -p miden-processor --lib` (1468 pass), processor
`trace::tests::lookup` (3 pass), `bus_degree_inventory` test passes.
