+++ air/src/constraints/lookup/buses/hash_kernel.rs
@@ -34,12 +34,10 @@ use miden_core::field::PrimeCharacteristicRing;
 use crate::{
     constraints::{
         logup_msg::{MemoryHeader, MemoryMsg, RangeMsg, SiblingMsgBitOne, SiblingMsgBitZero},
-        lookup::{
-            Deg, LookupBatch, LookupColumn, LookupGroup,
-            chiplet_air::{ChipletBusContext, ChipletLookupBuilder},
-        },
+        lookup::chiplet_air::{ChipletBusContext, ChipletLookupBuilder},
         utils::BoolNot,
     },
+    lookup::{Deg, LookupBatch, LookupColumn, LookupGroup},
     trace::{
         CHIPLETS_OFFSET,
         chiplets::{
