+++ air/src/constraints/lookup/debug/mod.rs
@@ -1,41 +1,23 @@
-//! Debug surface for the LogUp lookup-argument API.
+//! Miden-side debug surface for the LogUp lookup-argument API.
 //!
-//! Split into two regimes:
-//!
-//! | Module | Regime |
-//! |--------|--------|
-//! | [`validation`] | AIR self-checks — run against the `LookupAir` itself, no execution trace needed. Bundles encoding equivalence, scope, symbolic degree, and constants consistency via [`validation::validate`]. |
-//! | [`trace`] | Concrete-trace debugging — balance accumulator + per-column `(U, V)` oracle folds + mutex checks over a real main trace. |
-//!
-//! Entry points (all preserved from the previous layout):
-//!
-//! - [`validation::validate`] — bundled AIR self-check, returns one
-//!   [`validation::ValidationReport`].
-//! - [`inspect_structure`] / [`collect_inventory`] / [`check_encoding_equivalence`] /
-//!   [`check_challenge_scoping`] / [`check_symbolic_degrees`] — the individual validation pieces,
-//!   kept as wrappers so staged callers keep compiling.
-//! - [`check_trace_balance`] / [`collect_column_oracle_folds`] — trace-side entry points.
+//! Re-exports the generic debug surface from [`crate::lookup::debug`] and layers the
+//! Miden-specific symbolic degree-budget pass + composed [`validation::validate`] entry
+//! point on top.
 
-pub mod trace;
 pub mod validation;
 
-// Preserve the `debug::equivalence::...` and `debug::oracle::...` paths that the staged
-// `lookup/mod.rs` re-exports from. Aliases point at the consolidated builders.
-pub mod equivalence {
-    pub use super::validation::{DebugStructureBuilder as EquivalenceChecker, GroupMismatch};
-}
-pub mod oracle {
-    pub use super::trace::{DebugTraceBuilder as ColumnOracleBuilder, collect_column_oracle_folds};
-}
-
-pub use trace::{
-    BalanceReport, DebugTraceBuilder, MutualExclusionViolation, Unmatched, check_trace_balance,
-    collect_column_oracle_folds,
-};
+// Re-export the generic debug surface (structure records + entry points + trace-side
+// balance/oracle checks). Miden-side callers import from here so staged `lookup/mod.rs`
+// paths keep resolving.
 pub use validation::{
-    ColumnRecord, DEGREE_BUDGET, DebugStructure, DebugStructureBuilder, DegreeMismatch,
-    DegreeReport, EncodingMode, GroupMismatch, GroupRecord, InteractionRecord, Inventory, MultSign,
-    NumColumnsCheck, ScopeReport, ValidationReport, check_challenge_scoping,
-    check_encoding_equivalence, check_symbolic_degrees, collect_inventory, inspect_structure,
-    validate,
+    DEGREE_BUDGET, DegreeMismatch, DegreeReport, ValidationReport, check_symbolic_degrees, validate,
+};
+
+pub use crate::lookup::debug::{
+    BalanceReport, ColumnRecord, DebugStructure, DebugStructureBuilder, DebugTraceBuilder,
+    EncodingMode, GroupMismatch, GroupRecord, InteractionRecord, Inventory, MultSign,
+    MutualExclusionViolation, NumColumnsCheck, ScopeReport, StructureReport, Unmatched,
+    check_challenge_scoping, check_encoding_equivalence, check_trace_balance,
+    collect_column_oracle_folds, collect_inventory, equivalence, inspect_structure, oracle,
+    validate_structure_only,
 };
