+++ processor/src/trace/tests/lookup.rs
@@ -19,7 +19,7 @@
 use miden_air::{
     LOGUP_AUX_TRACE_WIDTH, LiftedAir, ProcessorAir,
     lookup::{
-        LookupChallenges, MIDEN_MAX_MESSAGE_WIDTH, MidenLookupAir, NUM_BUS_IDS, accumulate,
+        Challenges, MIDEN_MAX_MESSAGE_WIDTH, MidenLookupAir, NUM_BUS_IDS, accumulate,
         accumulate_slow, build_lookup_fractions, debug::collect_column_oracle_folds,
     },
 };
@@ -60,8 +60,7 @@ fn build_lookup_fractions_on_tiny_span() {
     let alpha = QuadFelt::new([raw[0], raw[1]]);
     let beta = QuadFelt::new([raw[2], raw[3]]);
     let air = MidenLookupAir;
-    let challenges =
-        LookupChallenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
+    let challenges = Challenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
 
     let fractions = build_lookup_fractions(&air, &main_trace, &periodic, &public_vals, &challenges);
 
@@ -132,8 +131,7 @@ fn build_lookup_fractions_matches_constraint_path_oracle() {
     let beta = QuadFelt::new([raw[2], raw[3]]);
     // --- Prover path: collect fractions and run the fused accumulator. ---
     let air = MidenLookupAir;
-    let challenges =
-        LookupChallenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
+    let challenges = Challenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
     let fractions = build_lookup_fractions(&air, &main_trace, &periodic, &public_vals, &challenges);
     // `accumulate` returns a row-major matrix with `num_rows + 1` rows and `num_cols`
     // columns; row 0 is the zero initial condition and row `r + 1` column `c` holds the
@@ -204,8 +202,7 @@ fn diagnostic_assembler_path_terminals() {
     let alpha = QuadFelt::new([raw[0], raw[1]]);
     let beta = QuadFelt::new([raw[2], raw[3]]);
     let air = MidenLookupAir;
-    let challenges =
-        LookupChallenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
+    let challenges = Challenges::<QuadFelt>::new(alpha, beta, MIDEN_MAX_MESSAGE_WIDTH, NUM_BUS_IDS);
     let fractions = build_lookup_fractions(&air, &main_trace, &periodic, &public_vals, &challenges);
     let aux = accumulate_slow(&fractions);
     let num_rows = trace.main_trace().num_rows();
